Update coding standard

+++ load/ret_3mth.sql
@@ -12,7 +12,6 @@ RETURNS VOID AS $$
 
 */
 
-
 DECLARE
 
   ref   RECORD;
@@ -26,6 +25,8 @@ DECLARE
 
   loop_counter INTEGER;
 
+  debug CHAR(3) := 'ON';
+
 BEGIN
   -- initialize.
   loop_counter := 0;
@@ -33,10 +34,14 @@ BEGIN
 
   -- date_start := (current_date - interval '3 month');
   SELECT start_date INTO date_start FROM lkp_start_date;
-  
-  raise notice 'Portfolio Method: ret_3mth'; 
-  raise notice 'Duration Type: %', dur_type;
-  raise notice 'Duration Period: %', stocks_to_choose;
+
+  IF debug = 'ON' THEN
+    raise notice 'Portfolio Method: ret_3mth'; 
+    raise notice 'Duration Type: %', dur_type;
+    raise notice 'Duration Period: %', stocks_to_choose;
+    raise notice 'date_start: %', date_start;
+    raise notice 'date_stop: %', date_stop;
+  END IF; 
 
   DELETE FROM summary 
   WHERE 1=1
@@ -44,7 +49,7 @@ BEGIN
     AND ret_type = dur_type 
     AND ret_period::integer = stocks_to_choose;
 
-	FOR ref IN SELECT * FROM lkp_dates WHERE prev_date BETWEEN date_start AND date_stop ORDER BY 1 LOOP
+	FOR ref IN SELECT * FROM lkp_dates WHERE lkp_prev_date BETWEEN date_start AND date_stop ORDER BY 1 LOOP
 
     loop_counter := loop_counter + 1;
 
@@ -55,15 +60,15 @@ BEGIN
       start_of_month := 'false';
     END IF;
 
-    IF (EXTRACT(MONTH FROM ref.data_date)) > (EXTRACT(MONTH FROM ref.prev_date)) OR 
-       (EXTRACT(YEAR FROM ref.data_date))  > (EXTRACT(YEAR FROM ref.prev_date)) 
+    IF (EXTRACT(MONTH FROM ref.lkp_data_date)) > (EXTRACT(MONTH FROM ref.lkp_prev_date)) OR 
+       (EXTRACT(YEAR FROM ref.lkp_data_date))  > (EXTRACT(YEAR FROM ref.lkp_prev_date)) 
     THEN
       start_of_month := 'true';
     ELSE 
       start_of_month := 'false';
     END IF;
 
-    IF ((EXTRACT(MONTH FROM ref.data_date) IN (2,4,6,8,10,12)) AND start_of_month = 'true')
+    IF ((EXTRACT(MONTH FROM ref.lkp_data_date) IN (2,4,6,8,10,12)) AND start_of_month = 'true')
     THEN
       start_of_two_month := 'true';
     ELSE
@@ -71,13 +76,15 @@ BEGIN
     END IF;
 
     raise notice '------------------';
-    raise notice 'Current date: %', ref.data_date;
-    raise notice 'Previous date: %', ref.prev_date;
+    raise notice 'Duration Type:  %', dur_type;
+    raise notice 'Stocks in list: %', stocks_to_choose;
+    raise notice 'Current date:   %', ref.lkp_data_date;
+    raise notice 'Previous date:  %', ref.lkp_prev_date;
     raise notice 'Start of month: %', start_of_month;
 
 
 		-- Find day_of_week.
-    day_of_week = date_part('dow', ref.prev_date);
+    day_of_week = date_part('dow', ref.lkp_prev_date);
 
     IF (dur_type = '2month' AND start_of_two_month)
     THEN
@@ -90,7 +97,7 @@ BEGIN
         SELECT symbol 
         FROM barchart_data 
         WHERE 1=1
-              AND data_date = ref.prev_date
+              AND data_date = ref.lkp_prev_date
       ORDER BY perc_chg_3mth::decimal DESC LIMIT stocks_to_choose;
 
     END IF;
@@ -104,7 +111,7 @@ BEGIN
         SELECT symbol 
         FROM barchart_data 
         WHERE 1=1
-              AND data_date = ref.prev_date
+              AND data_date = ref.lkp_prev_date
       ORDER BY perc_chg_3mth::decimal DESC LIMIT stocks_to_choose;
 
     END IF;
@@ -126,7 +133,7 @@ BEGIN
         SELECT symbol 
         FROM barchart_data 
         WHERE 1=1
-          AND data_date = ref.prev_date
+          AND data_date = ref.lkp_prev_date
         ORDER BY perc_chg_3mth::decimal DESC LIMIT stocks_to_choose;
 
         fortnight_counter := 0;
@@ -142,7 +149,7 @@ BEGIN
   		SELECT symbol 
   		FROM barchart_data 
   		WHERE 1=1
-        AND data_date = ref.prev_date
+        AND data_date = ref.lkp_prev_date
       ORDER BY perc_chg_3mth::decimal DESC LIMIT stocks_to_choose;
 
     END IF;
@@ -157,7 +164,7 @@ BEGIN
   			SELECT symbol 
   			FROM barchart_data 
   			WHERE 1=1
-  	  	  	  AND data_date = ref.prev_date
+  	  	  	  AND data_date = ref.lkp_prev_date
 			ORDER BY perc_chg_3mth::decimal DESC LIMIT stocks_to_choose;
 
 		END IF;
@@ -170,7 +177,7 @@ BEGIN
   	SELECT b.symbol, b.data_date, '3-mth', dur_type, stocks_to_choose, b.perc_change_daily 
 	  FROM barchart_data b 
   	WHERE 1=1
-  	  AND b.data_date = ref.data_date 
+  	  AND b.data_date = ref.lkp_data_date 
   	  AND b.symbol IN (SELECT symbol FROM tmp_stocks_to_invest);
 
 	-- raise notice '==============';
